backend/pyproject.toml
[project]
name = "agentic-rag-backend"
version = "1.0.0"
description = "Multi-step agentic RAG pipeline with Groq"
readme = "../README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=2.19.0,<4.0.0",
    "faiss-cpu>=1.9.0,<2.0.0",
    "fastapi>=0.111.0,<1.0.0",
    "langchain-community>=0.3.0,<1.0.0",
    "langchain>=0.3.0,<1.0.0",
    "langgraph>=0.2.0,<1.0.0",
    "numpy>=2.0.0,<3.0.0",
    "openai>=2.21.0",
    "pydantic>=2.7.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "ragas>=0.2.0,<1.0.0",
    "sentence-transformers>=3.0.0,<4.0.0",
    "uvicorn[standard]>=0.29.0,<1.0.0",
    "wikipedia>=1.4.0,<2.0.0",
]

[tool.uv]
package = false
